EICNET-1860: Make use of a base class for group_lfex based fields.
